+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ### (Work in Progress) Version 4.2.0
 
+#### Enterprise Policies
+
+- new policy to change the interface language of Firefox (Firefox 64+, Firefox ESR 60.3.1+), fixes
+  [#52](https://github.com/cadeyrn/enterprise-policy-generator/issues/52)
+
 #### Dependencies
 
 - updated eslint from version 5.9.0 to 5.10.0